mapToInt -> mapToLong 버그 수정

+++ src/main/java/com/ll/basic1/boundedContext/member/controller/MemberController.java
@@ -68,7 +68,7 @@ public class MemberController {
                     .filter(cookie ->
                             cookie.getName().equals("loginedMemberId"))
                     .map(Cookie::getValue)
-                    .mapToInt(Integer::parseInt)
+                    .mapToLong(Long::parseLong)
                     .findFirst()
                     .orElse(0);
         }
